% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{read_docx}
\alias{read_docx}
\title{Read data from a Modern Word File}
\usage{
read_docx(docx)
}
\arguments{
\item{docx}{The .docx file to read}
}
\value{
Vector of document text
}
\description{
Read data from a Modern Word File
}
\details{
Only accepts one file at a time and only .docx files.  Modifying file extensions will not work.

Text is broken out into the XML defined paragraphs in the vector.
}
\examples{
read_docx(docx = system.file('extdata','example.docx',package='readOffice'))

}
