% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{read_pptx}
\alias{read_pptx}
\title{Read data from a Modern PowerPoint File}
\usage{
read_pptx(pptx)
}
\arguments{
\item{pptx}{The .pptx file to read}
}
\value{
Vector of text, broken by paragraph
}
\description{
Read data from a Modern PowerPoint File
}
\details{
Only accepts one file at a time and only .pptx files.  Modifying file extensions will not work.
}
\examples{
read_pptx(system.file('extdata','example.pptx',package='readOffice'))

}

